requirements.txt
aiogram~=3.0
python-dotenv
betterlogging
google-generativeai
Pillow
